+++ tf_record_parser.py
@@ -140,7 +140,8 @@ class GGraphMutagenesisTFRecordProcessor:
             features["adj"] = self._create_float_feature(adj.astype(np.float32))
             features["node_feat"] = self._create_int_feature(node_feat)
             features["node_mask"] = self._create_float_feature(node_mask)
-            features["label"] = self._create_float_feature(label.astype(np.float32))
+            # features["label"] = self._create_float_feature(label.astype(np.float32))
+            features["label"] = self._create_float_feature(label)
 
             tf_example = tf.train.Example(
                 features=tf.train.Features(feature=features)
